+++ CDOWin/ViewModels/CreatePlacementViewModel.cs
@@ -25,7 +25,7 @@ public partial class CreatePlacementViewModel(IPlacementService service, DataInv
     public string? Position { get; set; }
     public DateTime? HireDate { get; set; }
     public DateTime? EndDate { get; set; }
-    public float? DaysOnJob { get; set; }
+    public int? DaysOnJob { get; set; }
     public string? Day1 { get; set; }
     public string? Day2 { get; set; }
     public string? Day3 { get; set; }
